pyproject.toml
[tool]
[tool.poetry]
name = "app"
version = "0.1.0"
description = ""
authors = [ "Marcus Schiesser <mail@marcusschiesser.de>" ]
readme = "README.md"

[tool.poetry.scripts]
generate = "generate:generate_index_to_pgvector_via_indexing_task"
"generate:pgvector_index" = "generate:generate_index_to_pgvector_via_indexing_task"
"generate:ui" = "generate:generate_ui_for_workflow"
"query" = "query:main"
dev = "main:run('dev')"
prod = "main:run('prod')"
evaluate = "rag.evaluate:run"

[tool.poetry.dependencies]
python = ">=3.11,<3.14"
python-dotenv = "^1.0.0"
aiostream = "==0.5.2"
celery = "==5.5.1"
boto3 = "==1.38.2"
llama-index-core = "==0.12.28"
llama-index-server = "==0.1.14"
llama-index-retrievers-bm25 = "==0.5.2"
llama-index-postprocessor-presidio = "==0.3.0"

[tool.poetry.group.llama-index-stores.dependencies]
llama-index-vector-stores-postgres = "==0.4.2"
llama-index-storage-docstore-redis = "==0.3.0"
llama-index-storage-index-store-redis = "==0.4.0"

[tool.poetry.group.llama-index-readers.dependencies]
llama-index-readers-web = "==0.3.9"
llama-index-readers-confluence = "==0.3.1"
llama-index-readers-database = "==0.3.0"
llama-index-readers-github = "==0.6.0"
llama-index-readers-slack = "==0.3.1"

[tool.poetry.dependencies.docx2txt]
version = "^0.8"

[tool.poetry.dependencies.llama-index-llms-openai]
version = "^0.3.2"

[tool.poetry.dependencies.llama-index-embeddings-openai]
version = "^0.3.1"

[tool.poetry.dependencies.llama-index-agent-openai]
version = "^0.4.0"

[tool.poetry.group]
[tool.poetry.group.dev]
[tool.poetry.group.dev.dependencies]
mypy = "^1.8.0"
pytest = "^8.3.5"
pytest-asyncio = "^0.25.3"

[tool.mypy]
python_version = "3.11"
plugins = "pydantic.mypy"
exclude = [ "tests", "venv", ".venv", "output", "config" ]
check_untyped_defs = true
warn_unused_ignores = false
show_error_codes = true
namespace_packages = true
ignore_missing_imports = true
follow_imports = "silent"
implicit_optional = true
strict_optional = false
disable_error_code = [ "return-value", "assignment" ]

[[tool.mypy.overrides]]
module = "app.*"
ignore_missing_imports = false

[build-system]
requires = [ "poetry-core" ]
build-backend = "poetry.core.masonry.api"
